+++ packages/solana-wallet-snap/src/core/services/signer/mocks/scenarios/types.ts
@@ -1,15 +1,21 @@
-import type { CompilableTransactionMessage } from '@solana/kit';
+import type { ExtendedKeyringAccount } from '@metamask/snap-networks-utils';
+import type {
+  TransactionMessage,
+  TransactionMessageWithBlockhashLifetime,
+  TransactionMessageWithFeePayer,
+} from '@solana/kit';
 
-import type { SolanaKeyringAccount } from '../../../../../entities';
 import type { Network } from '../../../../constants/solana';
 
 export type MockExecutionScenario = {
   name: string;
   scope: Network;
-  fromAccount: SolanaKeyringAccount;
-  toAccount: SolanaKeyringAccount;
+  fromAccount: ExtendedKeyringAccount;
+  toAccount: ExtendedKeyringAccount;
   fromAccountPrivateKeyBytes: Uint8Array;
-  transactionMessage: CompilableTransactionMessage;
+  transactionMessage: TransactionMessage &
+    TransactionMessageWithFeePayer &
+    TransactionMessageWithBlockhashLifetime;
   transactionMessageBase64Encoded: string;
   signedTransaction: any;
   signedTransactionBase64Encoded: string;