+++ packages/solana-wallet-snap/src/core/services/wallet/WalletService.ts
@@ -1,11 +1,15 @@
 import { SolMethod } from '@metamask/keyring-api';
-import type { Logger } from '@metamask/snap-networks-utils';
+import type {
+  ExtendedKeyringAccount,
+  Logger,
+} from '@metamask/snap-networks-utils';
 import type { Infer } from '@metamask/superstruct';
 import { assert, instance, object } from '@metamask/superstruct';
 import type { Commitment, SignatureBytes } from '@solana/kit';
 import {
   address as asAddress,
-  assertTransactionIsFullySigned,
+  assertIsFullySignedTransaction,
+  assertIsSendableTransaction,
   createKeyPairSignerFromPrivateKeyBytes,
   createSignableMessage,
   getBase58Codec,
@@ -17,7 +21,6 @@ import {
   verifySignature,
 } from '@solana/kit';
 
-import type { SolanaKeyringAccount } from '../../../entities';
 import { METAMASK_ORIGIN } from '../../constants/solana';
 import type { Caip10Address, Network } from '../../constants/solana';
 import type { DecompileTransactionMessageFetchingLookupTablesConfig } from '../../sdk-extensions/codecs';
@@ -90,7 +93,7 @@ export class WalletService {
    * @throws If the request is invalid.
    */
   async resolveAccountAddress(
-    keyringAccounts: SolanaKeyringAccount[],
+    keyringAccounts: ExtendedKeyringAccount[],
     scope: Network,
     request: SolanaWalletRequest,
   ): Promise<Caip10Address> {
@@ -161,7 +164,7 @@ export class WalletService {
    * @returns A Promise that resolves to the signed transaction.
    */
   async signTransaction(
-    account: SolanaKeyringAccount,
+    account: ExtendedKeyringAccount,
     transaction: string,
     scope: Network,
     origin: string,
@@ -204,7 +207,7 @@ export class WalletService {
 
     // If the transaction is fully signed, we can monitor it.
     try {
-      assertTransactionIsFullySigned(partiallySignedTransaction);
+      assertIsFullySignedTransaction(partiallySignedTransaction);
       const signature = getSignatureFromTransaction(partiallySignedTransaction);
       await this.#signatureMonitor.monitor(
         signature,
@@ -240,7 +243,7 @@ export class WalletService {
    * @returns A Promise that resolves to the signed transaction.
    */
   async signAndSendTransaction(
-    account: SolanaKeyringAccount,
+    account: ExtendedKeyringAccount,
     transactionMessageBase64Encoded: string,
     scope: Network,
     origin: string,
@@ -278,7 +281,7 @@ export class WalletService {
     const explorerUrl = getSolanaExplorerUrl(scope, 'tx', signature);
     this.#logger.info(`Sending transaction: ${explorerUrl}`);
 
-    assertTransactionIsFullySigned(partiallySignedTransaction);
+    assertIsSendableTransaction(partiallySignedTransaction);
 
     const sendConfig = {
       ...(options?.preflightCommitment
@@ -337,7 +340,7 @@ export class WalletService {
    * @returns A Promise that resolves to the signed message.
    */
   async signMessage(
-    account: SolanaKeyringAccount,
+    account: ExtendedKeyringAccount,
     message: string,
   ): Promise<SolanaSignMessageResponse> {
     this.#logger.log('Signing message', account, message);
@@ -392,7 +395,7 @@ export class WalletService {
    * @throws If the request is invalid.
    */
   async signIn(
-    account: SolanaKeyringAccount,
+    account: ExtendedKeyringAccount,
     params: SolanaSignInRequest['params'],
   ): Promise<SolanaSignInResponse> {
     this.#logger.log('Signing in', account, params);
@@ -427,7 +430,7 @@ export class WalletService {
    * signature is valid.
    */
   async verifySignature(
-    account: SolanaKeyringAccount,
+    account: ExtendedKeyringAccount,
     signatureBase58: Infer<typeof Base58Struct>,
     messageBase64: Infer<typeof Base64Struct>,
   ): Promise<boolean> {