+++ packages/stellar-wallet-snap/src/services/asset-metadata/AssetMetadataRepository.ts
@@ -1,3 +1,4 @@
+import type { IStateManager } from '@metamask/snap-networks-utils';
 import { cloneDeep } from 'lodash';
 
 import type {
@@ -5,7 +6,6 @@ import type {
   KnownCaip19AssetId,
   KnownCaip2ChainId,
 } from '../../api';
-import type { IStateManager } from '../state/IStateManager';
 import type {
   AssetMetadataByAssetId,
   AssetMetadataState,