+++ packages/stellar-wallet-snap/src/utils/__mocks__/snap.ts
@@ -17,9 +17,6 @@ export const getDefaultEntropySource = jest.fn();
 export const trackEvent = jest.fn();
 
 export const {
-  getState,
-  setState,
-  updateState,
   createInterface,
   showDialog,
   getPreferences,